Moved TIME_ARGS function definition to __init__.py
